+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0 (2025-01-XX)
 
 - feat: add view to display the list of installed Quarto extensions.
+- fix: update and correct trust authors and confirm installations prompts option value, *i.e.*, `Yes, never ask again`.
 - refactor(utils/extensions.ts): externalise user prompts to a separate module (`utils/ask.ts`).
 - refactor(extension.ts): don't use temporary variables for commands.